webkit: Compile JavaScriptCore as traditional ARM.
The JIT uses ARM assembly, and is configured to support thumb2
(v7+) or traditional (v4+) ARM instructions.  Without this,
Android will always compile the library with compiler set to use
the thumb instruction set, but the code will be traditional ARM.
In lieu of a thumb "v1" bit, just compile it as traditional ARM.

In an ideal world this would be conditional on the CPU type being
pre-v7.